+++ internal/domain/game/summary.go
@@ -5,6 +5,6 @@ import uuid "github.com/satori/go.uuid"
 type Summary struct {
 	SysID        uuid.UUID
 	PartyID      string
-	Seed         int64
+	Seed         int
 	Participants []*PlayerParticipant
 }
